gdb: Optimize section map
Store section map in an interval_tree.

Remove all the code that deals with lazy updates and dangling pointers,
which result from them, since the updates are very fast now.

Make sure updates are performed after sym_offsets calls, otherwise
section addresses are either not available or not up-to-date.

Keep filtering debuginfo and overlapping sections, but do that on
lookup rather than on insertion.  The reason is that we want to keep
them in case one of the two overlapping sections goes away.