gdb.perf/: Allow adjusting GenPerfTest compile options
Introduce new callbacks to adjust the compile options based on
certain criteria: the total number of shared objects for the binary and
the shared object index for shared objects (no criteria are defined for
the tail shared object at the moment).  This is useful for setting
shared objects' load addresses in performance tests.